added all files test
