+++ client/src/test/2_stress.test.ts
@@ -1,5 +1,7 @@
 import assert from 'assert';
-import TestHelper, { CARBON_NAME, SETUP_TIMEOUT, SILICON_NAME, SIMPLE } from './TestHelper';
+import { readdir } from 'fs/promises';
+import { Helper } from '../Helper';
+import TestHelper, { CARBON_NAME, DATA_ROOT, LONG, EMPTY, SETUP_TIMEOUT, SILICON_NAME, SIMPLE } from './TestHelper';
 
 suite('ViperIDE Stress Tests', () => {
 
@@ -75,15 +77,35 @@ suite('ViperIDE Stress Tests', () => {
         assert(!TestHelper.hasObservedInternalError());
     });
 
-    test("Test simple verification with carbon", async function() {
-        this.timeout(35000);
+    test("5. rapidly switch between files without waiting for verification", async function() {
+        this.timeout(10000);
+
+        TestHelper.resetErrors();
+
+        const files = [SIMPLE, LONG, EMPTY, SIMPLE, LONG, EMPTY, SIMPLE, LONG, EMPTY];
+        for (const file of files) {
+            await TestHelper.openFile(file);
+        }
 
         await TestHelper.openFile(SIMPLE);
-        const carbonVerified = TestHelper.waitForVerification(SIMPLE, CARBON_NAME);
-        await TestHelper.selectBackend(CARBON_NAME);
-        await carbonVerified;
-        const siliconVerified = TestHelper.waitForVerification(SIMPLE, SILICON_NAME);
-        await TestHelper.selectBackend(SILICON_NAME);
-        await siliconVerified;
+        await TestHelper.verify();
+        await TestHelper.waitForVerification(SIMPLE);
+        assert(!TestHelper.hasObservedInternalError(), "internal error detected while rapidly switching files");
     });
+
+    test("6. verify all files in workspace in quick succession", async function() {
+        this.timeout(100000);
+
+        TestHelper.resetErrors();
+
+        const allFiles = await readdir(DATA_ROOT);
+        const viperFiles = allFiles.filter(f => Helper.isViperSourceFile(f));
+        assert(viperFiles.length > 0, "no Viper files found in test data");
+
+        for (const fileName of viperFiles) {
+            await TestHelper.openAndVerify(fileName);
+        }
+        assert(!TestHelper.hasObservedInternalError(), "internal error detected while verifying files in quick succession");
+    });
+
 });
